Move IR matching classes into separate packages
