+++ test/hotspot/jtreg/compiler/lib/ir_framework/IRNode.java
@@ -23,7 +23,7 @@
 
 package compiler.lib.ir_framework;
 
-import compiler.lib.ir_framework.driver.IRMatcher;
+import compiler.lib.ir_framework.driver.irmatching.IRMatcher;
 import compiler.lib.ir_framework.shared.*;
 import jdk.test.lib.Platform;
 import sun.hotspot.WhiteBox;
