+++ test/hotspot/jtreg/compiler/lib/ir_framework/driver/irmatching/IRViolationException.java
@@ -21,7 +21,7 @@
  * questions.
  */
 
-package compiler.lib.ir_framework.driver;
+package compiler.lib.ir_framework.driver.irmatching;
 
 import compiler.lib.ir_framework.IR;
 import compiler.lib.ir_framework.Test;
