+++ test/hotspot/jtreg/compiler/lib/ir_framework/driver/irmatching/irrule/Counts.java
@@ -21,7 +21,7 @@
  * questions.
  */
 
-package compiler.lib.ir_framework.driver;
+package compiler.lib.ir_framework.driver.irmatching.irrule;
 
 import compiler.lib.ir_framework.IR;
 import compiler.lib.ir_framework.shared.Comparison;
