+++ test/hotspot/jtreg/compiler/lib/ir_framework/driver/irmatching/irrule/CountsMatchResult.java
@@ -21,30 +21,33 @@
  * questions.
  */
 
-package compiler.lib.ir_framework.driver;
+package compiler.lib.ir_framework.driver.irmatching.irrule;
 
 import compiler.lib.ir_framework.IR;
 
-import java.util.List;
+import java.util.ArrayList;
 
 /**
- * Base class representing a result of an applied check attribute of an IR rule.
+ * Class representing a result of an applied counts attribute of an IR rule.
  *
- * @see IR
+ * @see IR#counts()
  */
-abstract class CheckAttributeMatchResult implements MatchResult {
-    protected List<RegexFailure> regexFailures = null;
+class CountsMatchResult extends CheckAttributeMatchResult {
 
-    public int getMatchesCount() {
-        if (fail()) {
-            return regexFailures.stream().map(RegexFailure::getMatchesCount).reduce(0, Integer::sum);
-        } else {
-            return 0;
+    public void addFailure(RegexFailure regexFailure) {
+        if (regexFailures == null) {
+            regexFailures = new ArrayList<>();
         }
+        regexFailures.add(regexFailure);
     }
 
     @Override
-    public boolean fail() {
-        return regexFailures != null;
+    public String buildFailureMessage() {
+        StringBuilder failMsg = new StringBuilder();
+        failMsg.append("     - counts: Graph contains wrong number of nodes:").append(System.lineSeparator());
+        for (RegexFailure regexFailure : regexFailures) {
+            failMsg.append(regexFailure.getFormattedFailureMessage());
+        }
+        return failMsg.toString();
     }
 }
