+++ test/hotspot/jtreg/compiler/lib/ir_framework/driver/irmatching/irrule/FailOnRegexFailure.java
@@ -21,27 +21,28 @@
  * questions.
  */
 
-package compiler.lib.ir_framework.driver;
+package compiler.lib.ir_framework.driver.irmatching.irrule;
 
-import compiler.lib.ir_framework.IR;
-
-import java.util.ArrayList;
 import java.util.List;
-import java.util.regex.Matcher;
 
 /**
- * Base class representing a check attribute of an IR rule.
+ * This class represents an IR matching failure of a regex of a failOn attribute of an IR rule.
  *
- * @see IR
+ * @see FailOn
  */
-abstract class CheckAttribute {
-    protected List<String> getMatchedNodes(Matcher m) {
-        List<String> matches = new ArrayList<>();
-        do {
-            matches.add(m.group());
-        } while (m.find());
-        return matches;
+class FailOnRegexFailure extends RegexFailure {
+    public FailOnRegexFailure(String nodeRegex, int nodeId, List<String> matches) {
+        super(nodeRegex, nodeId, matches);
     }
 
-    abstract public CheckAttributeMatchResult apply(String compilation);
+    @Override
+    public String getFormattedFailureMessage() {
+        return getRegexLine()
+               + getMatchedNodesBlock();
+    }
+
+    @Override
+    protected String getMatchedPrefix() {
+        return "Matched forbidden";
+    }
 }
