+++ test/hotspot/jtreg/testlibrary_tests/ir_framework/examples/IRExample.java
@@ -24,7 +24,7 @@
 package ir_framework.examples;
 
 import compiler.lib.ir_framework.*;
-import compiler.lib.ir_framework.driver.IRViolationException;
+import compiler.lib.ir_framework.driver.irmatching.IRViolationException;
 
 /*
  * @test
