+++ test/hotspot/jtreg/testlibrary_tests/ir_framework/tests/TestIRMatching.java
@@ -24,7 +24,7 @@
 package ir_framework.tests;
 
 import compiler.lib.ir_framework.*;
-import compiler.lib.ir_framework.driver.IRViolationException;
+import compiler.lib.ir_framework.driver.irmatching.IRViolationException;
 import jdk.test.lib.Asserts;
 import jdk.test.lib.Platform;
 import sun.hotspot.WhiteBox;
